[+AM] Prevent cofiguration to crash on startup
Previously we had an issue with the order of when things are loaded. In most cases we start sapience before initializing dotenv and in these cases we would crash on startup.

The solution before this change was to load dotenv before the rails application like in `config/environment.rb`. For rake tasks we didn't load environment but directly went to loading `config/application.rb`.

It ended up being a lot of changes in a lot of applications and to remedy this we decided to postpone some validations to ensure we are not so dependent on when things are loaded. Sentry for instance will be using any `SENTRY_DSN` value when it is present and we will output a scheduled warning when it isn't present at all.

Statsd not being a requirement for a functioning app will be treated the same way while the stream appender will only check if the stream is open.