Decouple Appender::Sentry unit tests from Sapience.add_appender
There's no reason these tests should fail if someone breaks the
`add_appender` method, so I've modified them to just use a
`described_class.new` as the subject.

However, it's probably useful to make sure we don't break
Appender::Sentry in such a way that it can't be used by the
Sapience.add_appender method, so I've added in a test to cover this off.

There's definitely a case for moving that test out into a shared example
re-used between Appenders, happy to do so if anyone wants!