"Fix" heisenspec
If we log an "message" instead of an "exception" here (i.e. a Log
without an Exception), then this spec will fail if it's run after any of
the shared examples in the file.

This is due to the fact that in order to check we're setting
configuration correctly, we're using an `instance_spy` to set the `dsn`
and `tags` elements, but not *actually* changing the Configuration from
what gets configured in the previous specs. Due to some craziness in
Raven, this means we need to set the exception here when we log the
message.

FUN FACT: If you pull this test out of the `describe` block, it also
works. Despite having spent several hours investigating this, I have
absolutely no fucking idea why :)