Reimplementation of grape 404s
Grape expects us to throw(:error, {status: code, message: ""} when we have a
404. - Its a rails nom to raise exceptions for 404,
ActiveRecord::RecordNotFound being a notable example.

Sapience checks if a "NotFound" type error is raised, - it logs this with the
correct status code, and throws the error object that Grape expects. -
Previously Sapience logged this as a 500 error, and did not correctly
throw(:error) as expected by Grape.